XIVY-3256 Count user with new UserQuery API

Approved-by: Lukas Lieb
Approved-by: Reto Weiss

+++ engine-cockpit/src/ch/ivyteam/enginecockpit/ApplicationDetailBean.java
@@ -84,7 +84,7 @@ public class ApplicationDetailBean
     return getIApplication().getSecurityContext().getSessionCount();
   }
   
-  public int getUsersCount()
+  public long getUsersCount()
   {
     return security.getUsersCount();
   }
